Pull request by automated build

+++ card_driver/card/driver/sph_cs/sphcs_pcie.h
@@ -40,7 +40,7 @@ struct sphcs_dma_hw_ops {
 	int (*init_dma_engine)(void *hw_handle);
 	int (*init_lli)(void *hw_handle, struct lli_desc *outLli, struct sg_table *src, struct sg_table *dst, uint64_t dst_offset, bool single_list);
 	u64 (*gen_lli)(void *hw_handle, struct sg_table *src, struct sg_table *dst, struct lli_desc *outLli, uint64_t dst_offset);
-	int (*edit_lli)(void *hw_handle, struct lli_desc *outLli, uint32_t size);
+	int (*edit_lli)(void *hw_handle, struct lli_desc *outLli, uint64_t size);
 	int (*init_lli_vec)(void *hw_handle, struct lli_desc *outLli, uint64_t dst_offset, genlli_get_next_cb cb, void *cb_ctx);
 	u64 (*gen_lli_vec)(void *hw_handle, struct lli_desc *outLli, uint64_t dst_offset, genlli_get_next_cb cb, void *cb_ctx);
 	int (*edit_lli_elem)(struct lli_desc *lli, u32 elem_idx, dma_addr_t src, dma_addr_t dst);
